Fix minor bugs LOGIN and LOGGER modules.
Fixed some minor issues with the LOGIN and the LOGGER modules. Now, admins can no longer login into the second-years' login portal.

The Logger module was malfunctioning. This turned out to be because that the file mode was in 'w' rather than 'a' so it was overwriting the file each time.