Fix Admin-Login bug.
Fixed a bug where the admins or any user could go to a page in any of the admin-modules without even logging if he/she could log in.

Also fixed a bug where any user if he/she was logged in could view the admin pages and edit content. This was rectified and, now, all the pages can only be viewed only if an admin is logged in.